semantic fixes to SNMP-USER-BASED-SM_MIB
The semantics allow the usmUserAuthKeyChange and usmUserPrivKeyChange
objects to be written to in the same set requests that also creates and
clones the user. This was not possible beforehand, causing test tools
checking semantic SNMPv3 behaviour to fail on a lot of test cases.

Furthermore, once the user has been cloned by writing to an instance of
usmUserCloneFrom, further set-operations to the same object will not
return an error, but be no-ops. Especially, it must be avoided to copy
security parameters again (possibly even from a different user).