Errors in vacmAccessTable RowStatus handling
There are problems with the handling of vacmAccessTableStatus that cause
some SNMP test suites to report errors.

Most notably, errorneous set operations frequently cause "genErr" errors
to be returned which usually is a sign for bad coding. These "genErr"
errors are usually caused by badmatch exceptions coming from
  {ok, Row} = snmpa_vacm:get_row(RowIndex)
if the row does not exist.

This patch fixes the badmatch errors and adjusts the semantic of the
RowStatus handling in that table to be compliant with the RowStatus
textual description of SNPMv2-TC MIB.